way3sms
Send sms via a terminal. Uses python and way2sms.

The following code can be easily used to send sms for free via way2sms.
The advantage here is that the code when compiled using python asks
message and number to be sent to and automatically sends the message. No
need to login and click again and again. Here's it in action: